An attempt at cleaning up/consolidating some logic (its still pretty messy)
Uses a sort-of-event system to load seeds once and then have multiple listeners (DataAggregator's) that read and dump the relevant data to multiple different csv files.

Each seed still takes a bit of time and there's definitely memory being leaked that needs to be fixed (my C++ knowledge is insanely rusty and way out-of-date).